Removed Random.org dep and moved to /manager group
