+++ commands/manager_commands.py
@@ -4,12 +4,18 @@ from discord import Object
 from config import Config
 import enum
 import logging
-
+import random
 
 class VODType(enum.Enum):
     approved = 1
     rejected = 2
     complete = 3
+
+FIRST_HALF_STARTING_ROUNDS = [1,2,3]
+FIRST_HALF_NORMAL_ROUNDS_START = 4
+FIRST_HALF_NORMAL_ROUNDS_END = 12
+SECOND_HALF_STARTING_ROUNDS = [13,14,15]
+SECOND_HALF_NORMAL_ROUNDS_START = 16
 
 
 @app_commands.guild_only()
@@ -43,3 +49,27 @@ class ManagerCommands(app_commands.Group, name="manager"):
 
         await interaction.channel.add_tags(Object(id=tag_id))
         await interaction.response.send_message("Applied tag.", ephemeral=True)
+
+    @app_commands.command(name="get_review_rounds")
+    @app_commands.checks.has_role("VOD Review Team")
+    @app_commands.describe(rounds="total number of rounds in VOD")
+    async def get_rounds(
+        self, interaction: Interaction, rounds: int
+    ) -> None:
+        """Generates rounds to check for the pre-round comms requirement"""
+        if (rounds < 21):
+            await interaction.response.send_message("Not enough rounds in VOD. Match must be 13-8 or closer.\n;rejectedforfinalscore", ephemeral=True)
+            return
+        roundsToCheck = []
+        returnString = "Pre-round Comms:"
+        random.shuffle(FIRST_HALF_STARTING_ROUNDS) # Get two rounds from the first 3 rounds of each half
+        random.shuffle(SECOND_HALF_STARTING_ROUNDS)
+        roundsToCheck += FIRST_HALF_STARTING_ROUNDS[:2]
+        roundsToCheck += SECOND_HALF_STARTING_ROUNDS[:2]
+        roundsToCheck.append(random.randint(FIRST_HALF_NORMAL_ROUNDS_START, FIRST_HALF_NORMAL_ROUNDS_END)) # Add one more non-start round from each half
+        roundsToCheck.append(random.randint(SECOND_HALF_NORMAL_ROUNDS_START, rounds)) # Second round ends at total round number
+        roundsToCheck.sort()
+        for num in roundsToCheck:
+            returnString += f"\nRound {num}:" # Generates response that VOD Reviewer can copy-paste into the forum 
+
+        await interaction.response.send_message(returnString, ephemeral=True)
