+++ config.ini.example
@@ -35,6 +35,3 @@ Username =
 Password =
 Host =
 Name =
-
-[RandomOrg]
-ApiKey =
