+++ requirements.txt
@@ -13,5 +13,3 @@ quart==0.18.4
 quart_cors==0.6.0
 hypercorn==0.14.4
 pytimeparse==1.1.8
-rdoclient==1.5.0
-six==1.16.0
